fix all pylint; cleanup apis and tests.
