+++ nums/core/array/base.py
@@ -375,4 +375,3 @@ class BlockArrayBase(object):
             broadcast = it.itviews[0]
         result.blocks = broadcast
         return result
-
