+++ nums/core/systems/numpy_compute.py
@@ -15,7 +15,6 @@
 
 
 import random
-import secrets
 import numpy as np
 from numpy.random import PCG64
 from numpy.random import Generator
