+++ nums/numpy/fft.py
@@ -14,4 +14,7 @@
 # limitations under the License.
 
 
+# pylint: disable = unused-import
+
+
 from nums.core.application_manager import instance as _instance
