+++ pylintrc
@@ -40,6 +40,7 @@ disable=R,
         consider-using-enumerate,
         unnecessary-lambda,
         misplaced-comparison-constant,
+        unsubscriptable-object,
 #         redefined-builtin,
 #         protected-name,
 #         no-else-return,
