+++ tests/numpy/test_np_linalg.py
@@ -14,11 +14,14 @@
 # limitations under the License.
 
 
-from nums.numpy import BlockArray
+# pylint: disable=import-outside-toplevel
 
 
 def test_basic(nps_app_inst):
     import nums.numpy as nps
+
+    assert nps_app_inst is not None
+
     x = nps.random.RandomState(1337).random_sample((100, 10))
     q, r = nps.linalg.qr(x)
     assert nps.allclose(x, q @ r)
